Change project name in get started tutorial
